requirements.txt
Flask==3.0.3
Gunicorn
Markdown==3.4.4
google-genai==1.24.0
Pillow==9.5.0
protobuf
python-dotenv==1.0.1

